+++ packages/livekit-rtc/src/index.ts
@@ -17,8 +17,8 @@ export {
 } from './track.js';
 export { VideoFrame } from './video_frame.js';
 export { AudioFrame } from './audio_frame.js';
-export { AudioStream, AudioStreamEvent, AudioFrameEvent } from './audio_stream.js';
-export { VideoStream, VideoStreamEvent, VideoFrameEvent } from './video_stream.js';
+export { AudioStream } from './audio_stream.js';
+export { VideoStream, VideoFrameEvent } from './video_stream.js';
 export { AudioSource } from './audio_source.js';
 export { VideoSource } from './video_source.js';
 export {
@@ -40,6 +40,6 @@ export {
 export { RpcRequest, RpcAck, RpcResponse, RpcError } from './rpc.js';
 export { EncryptionType, EncryptionState } from './proto/e2ee_pb.js';
 export { StreamState, TrackKind, TrackSource } from './proto/track_pb.js';
-export { VideoBufferType, VideoRotation } from './proto/video_frame_pb.js';
+export { VideoBufferType, VideoRotation, VideoCodec } from './proto/video_frame_pb.js';
 export { ParticipantKind } from './proto/participant_pb.js';
 export { dispose } from './ffi_client.js';