+++ packages/livekit-rtc/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @livekit/rtc-node
 
+## 0.9.2
+
+### Patch Changes
+
+- expose VideoCodec - [#283](https://github.com/livekit/node-sdks/pull/283) ([@nbsp](https://github.com/nbsp))
+
+## 0.9.1
+
+### Patch Changes
+
+- Fix ms conversion bug in AudioSource - [#279](https://github.com/livekit/node-sdks/pull/279) ([@bcherry](https://github.com/bcherry))
+
+## 0.9.0
+
+### Minor Changes
+
+- make AudioStream use AsyncIterableIterator instead of typed EventEmitter - [#272](https://github.com/livekit/node-sdks/pull/272) ([@nbsp](https://github.com/nbsp))
+
+### Patch Changes
+
+- bump ffi to v0.10.2 - [#271](https://github.com/livekit/node-sdks/pull/271) ([@nbsp](https://github.com/nbsp))
+
 ## 0.8.1
 
 ### Patch Changes